test: rename mjs files to js files for codesandbox compatibility
