+++ .bin/color.js
