+++ .bin/internal/dirs.js
