+++ .bin/test-coverage.js
@@ -1,8 +1,8 @@
 import cp from 'child_process'
 import fp from 'path'
-import fs from './fs.mjs'
+import fs from './fs.js'
 import os from 'os'
-import { bold, dim, green, red } from './color.mjs'
+import { bold, dim, green, red } from './color.js'
 
 // Creates a unique temporary directory
 !(async () => {
