+++ .bin/test-watch.js
@@ -7,7 +7,7 @@ nodemon(
 		`--watch packages/core/tests`,
 		`--watch packages/react/src`,
 		`--watch packages/react/tests`,
-		`--exec "clear; ${['node', '.bin/test.mjs'].concat(process.argv.slice(2)).join(' ')}"`,
+		`--exec "clear; ${['node', '.bin/test.js'].concat(process.argv.slice(2)).join(' ')}"`,
 	].join(' '),
 )
 	.on('start', () => {
